+++ packages/utils/src/index.d.ts
@@ -7,7 +7,7 @@ interface String {
    *
    * @public
    */
-  equal(val: string | number): boolean
+  equal(val: string | number): boolean;
 
 
   /**
@@ -26,7 +26,15 @@ interface String {
    *
    * @public
    */
-  capitalize(): string
+  capitalize(): string;
+
+  /**
+   *
+   * 是否包含 containerValue 字符串
+   *
+   * @public
+   */
+  container(containerValue?: string): boolean;
 
 
   /**
@@ -41,7 +49,7 @@ interface String {
    * @public
    *
    */
-  escapeRegexp(): string
+  escapeRegexp(): string;
 }
 
 
@@ -54,7 +62,7 @@ interface Number {
    *
    * @public
    */
-  equal(val: string | number): boolean
+  equal(val: string | number): boolean;
 
 
   /**
@@ -67,4 +75,17 @@ interface Number {
    * @public
    */
   log(mark?: string): string | number;
+
+
+  /**
+   *
+   * 返回自动生成随机数 不会修改原来数字
+   *
+   * @param min - 最小数字
+   *
+   * @param max - 最大数字
+   *
+   * @public
+   */
+  random(min: number, max: number): number;
 }